requirements.txt
torch
torchvision
hydra-core
wandb
einops
diffusers
transformers